monitoring/metrics_collector.py
"""
Metrics Collector Implementation
"""

import psutil
import time
from datetime import datetime
from typing import List, Optional, Tuple
from .interfaces import SystemMetrics


class MetricsCollector:
    """Collector for system performance metrics."""
    
    def __init__(self):
        """Initialize the metrics collector."""
        self.metrics_history = []
    
    def collect_metrics(self) -> SystemMetrics:
        """Collect current system metrics."""
        try:
            # Get system metrics
            cpu_usage = psutil.cpu_percent()
            memory = psutil.virtual_memory()
            memory_usage = memory.percent
            
            # Mock federated learning metrics
            metrics = SystemMetrics(
                timestamp=datetime.now(),
                cpu_usage=cpu_usage,
                memory_usage=memory_usage,
                active_clients=0,  # Will be updated by system
                training_rounds_completed=0,
                anomalies_detected=0,
                authentication_failures=0
            )
            
            self.metrics_history.append(metrics)
            return metrics
            
        except Exception:
            # Fallback metrics if psutil not available
            return SystemMetrics(
                timestamp=datetime.now(),
                cpu_usage=50.0,
                memory_usage=60.0,
                active_clients=0,
                training_rounds_completed=0,
                anomalies_detected=0,
                authentication_failures=0
            )
    
    def record_metric(self, name: str, value: float, timestamp: Optional[datetime] = None) -> None:
        """Record a custom metric."""
        if timestamp is None:
            timestamp = datetime.now()
        
        # Store metric (simplified implementation)
        pass
    
    def get_metric_history(self, name: str, start_time: datetime, end_time: datetime) -> List[Tuple]:
        """Get historical values for a metric."""
        # Simplified implementation
        return []