pq_security/kyber.py
"""
CRYSTALS-Kyber Key Exchange Implementation

Provides quantum-resistant key exchange using CRYSTALS-Kyber algorithm.
Includes fallback simulation when pqcrypto library is unavailable.
"""

import os
import hashlib
import secrets
from typing import Tuple, Optional
from .interfaces import IKeyExchange


class KyberKeyExchange(IKeyExchange):
    """CRYSTALS-Kyber key exchange implementation with fallback simulation."""
    
    def __init__(self, security_level: int = 3):
        """Initialize Kyber key exchange.
        
        Args:
            security_level: Security level (2, 3, or 4 corresponding to Kyber512, Kyber768, Kyber1024)
        """
        self.security_level = security_level
        self._pqcrypto_available = self._check_pqcrypto_availability()
        
        # Security level parameters based on NIST specifications
        self._params = {
            2: {'n': 256, 'q': 3329, 'k': 2, 'eta1': 3, 'eta2': 2, 'du': 10, 'dv': 4},
            3: {'n': 256, 'q': 3329, 'k': 3, 'eta1': 2, 'eta2': 2, 'du': 10, 'dv': 4},
            4: {'n': 256, 'q': 3329, 'k': 4, 'eta1': 2, 'eta2': 2, 'du': 11, 'dv': 5}
        }
        
        if security_level not in self._params:
            raise ValueError(f"Unsupported security level: {security_level}")
    
    def _check_pqcrypto_availability(self) -> bool:
        """Check if pqcrypto library is available."""
        try:
            import pqcrypto.kem.kyber768
            return True
        except ImportError:
            return False
    
    def generate_keypair(self) -> Tuple[bytes, bytes]:
        """Generate Kyber keypair.
        
        Returns:
            Tuple of (public_key, private_key) as bytes
        """
        if self._pqcrypto_available:
            return self._generate_keypair_real()
        else:
            return self._generate_keypair_simulation()
    
    def encapsulate(self, public_key: bytes) -> Tuple[bytes, bytes]:
        """Encapsulate shared secret using public key.
        
        Args:
            public_key: Public key for encapsulation
            
        Returns:
            Tuple of (ciphertext, shared_secret)
        """
        if self._pqcrypto_available:
            return self._encapsulate_real(public_key)
        else:
            return self._encapsulate_simulation(public_key)
    
    def decapsulate(self, ciphertext: bytes, private_key: bytes) -> bytes:
        """Decapsulate shared secret from ciphertext.
        
        Args:
            ciphertext: Ciphertext containing encapsulated secret
            private_key: Private key for decapsulation
            
        Returns:
            Shared secret as bytes
        """
        if self._pqcrypto_available:
            return self._decapsulate_real(ciphertext, private_key)
        else:
            return self._decapsulate_simulation(ciphertext, private_key)
    
    def _generate_keypair_real(self) -> Tuple[bytes, bytes]:
        """Generate keypair using real pqcrypto library."""
        if self.security_level == 2:
            import pqcrypto.kem.kyber512
            return pqcrypto.kem.kyber512.keypair()
        elif self.security_level == 3:
            import pqcrypto.kem.kyber768
            return pqcrypto.kem.kyber768.keypair()
        elif self.security_level == 4:
            import pqcrypto.kem.kyber1024
            return pqcrypto.kem.kyber1024.keypair()
    
    def _encapsulate_real(self, public_key: bytes) -> Tuple[bytes, bytes]:
        """Encapsulate using real pqcrypto library."""
        if self.security_level == 2:
            import pqcrypto.kem.kyber512
            return pqcrypto.kem.kyber512.enc(public_key)
        elif self.security_level == 3:
            import pqcrypto.kem.kyber768
            return pqcrypto.kem.kyber768.enc(public_key)
        elif self.security_level == 4:
            import pqcrypto.kem.kyber1024
            return pqcrypto.kem.kyber1024.enc(public_key)
    
    def _decapsulate_real(self, ciphertext: bytes, private_key: bytes) -> bytes:
        """Decapsulate using real pqcrypto library."""
        if self.security_level == 2:
            import pqcrypto.kem.kyber512
            return pqcrypto.kem.kyber512.dec(ciphertext, private_key)
        elif self.security_level == 3:
            import pqcrypto.kem.kyber768
            return pqcrypto.kem.kyber768.dec(ciphertext, private_key)
        elif self.security_level == 4:
            import pqcrypto.kem.kyber1024
            return pqcrypto.kem.kyber1024.dec(ciphertext, private_key)
    
    def _generate_keypair_simulation(self) -> Tuple[bytes, bytes]:
        """Generate keypair using simulation based on NIST specifications."""
        params = self._params[self.security_level]
        
        # Generate a seed for deterministic key generation
        seed = secrets.token_bytes(32)
        
        # Derive public and private keys from seed
        public_key_size = params['k'] * 32 + 32  # Simplified calculation
        private_key_size = params['k'] * 32 + 64  # Don't include public key in private key
        
        # Generate public key from seed
        public_key_data = hashlib.sha3_256(seed + b"public").digest()
        while len(public_key_data) < public_key_size:
            public_key_data += hashlib.sha3_256(public_key_data).digest()
        public_key = public_key_data[:public_key_size]
        
        # Generate private key from seed (linked to public key)
        private_key_data = hashlib.sha3_256(seed + b"private").digest()
        while len(private_key_data) < private_key_size:
            private_key_data += hashlib.sha3_256(private_key_data).digest()
        
        # Store the seed in private key for decapsulation
        private_key = seed + private_key_data[:private_key_size - 32]
        
        return public_key, private_key
    
    def _encapsulate_simulation(self, public_key: bytes) -> Tuple[bytes, bytes]:
        """Encapsulate using simulation."""
        params = self._params[self.security_level]
        
        # Generate random shared secret (32 bytes for 256-bit security)
        shared_secret = secrets.token_bytes(32)
        
        # Simulate ciphertext size based on Kyber parameters
        ciphertext_size = params['k'] * params['du'] * 32 + params['dv'] * 32
        
        # Create encryption key from public key
        encryption_key = hashlib.sha3_256(public_key + b"encrypt").digest()
        
        # Encrypt shared secret using XOR (simple but deterministic)
        encrypted_secret = bytes(a ^ b for a, b in zip(shared_secret, encryption_key))
        
        # Create rest of ciphertext with random data
        remaining_size = ciphertext_size - 32
        if remaining_size > 0:
            # Use deterministic padding based on public key and shared secret
            padding_seed = hashlib.sha3_256(public_key + shared_secret).digest()
            padding_data = padding_seed
            while len(padding_data) < remaining_size:
                padding_data += hashlib.sha3_256(padding_data).digest()
            padding_data = padding_data[:remaining_size]
        else:
            padding_data = b""
        
        ciphertext = encrypted_secret + padding_data
        
        return ciphertext, shared_secret
    
    def _decapsulate_simulation(self, ciphertext: bytes, private_key: bytes) -> bytes:
        """Decapsulate using simulation."""
        if len(ciphertext) < 32:
            raise ValueError("Ciphertext too short")
        
        # Extract seed from private key (first 32 bytes)
        seed = private_key[:32]
        
        # Reconstruct public key from seed (same as in key generation)
        params = self._params[self.security_level]
        public_key_size = params['k'] * 32 + 32
        
        public_key_data = hashlib.sha3_256(seed + b"public").digest()
        while len(public_key_data) < public_key_size:
            public_key_data += hashlib.sha3_256(public_key_data).digest()
        public_key = public_key_data[:public_key_size]
        
        # Create decryption key (same as encryption key)
        decryption_key = hashlib.sha3_256(public_key + b"encrypt").digest()
        
        # Extract and decrypt the shared secret
        encrypted_secret = ciphertext[:32]
        shared_secret = bytes(a ^ b for a, b in zip(encrypted_secret, decryption_key))
        
        return shared_secret
    
    @property
    def is_simulation_mode(self) -> bool:
        """Check if running in simulation mode."""
        return not self._pqcrypto_available
    
    def get_security_level(self) -> int:
        """Get current security level."""
        return self.security_level
    
    def get_ciphertext_size(self) -> int:
        """Get expected ciphertext size for current security level."""
        params = self._params[self.security_level]
        # Ciphertext size: c1 (k * 320 bytes) + c2 (128 bytes)
        return params['k'] * 320 + 128